Increased supported numbers from 3981 to 3982.
My friend tried to use this project but the program returned inconclusive result for number he wanted to try: 3982, so by making minimal changes to codebase, following it's current code layout, I've made a change that allows to conclusively answer for number 3982. Of course this has been thoroughly tested and verified using state of the art unit tests.